+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- Only show the message timestamp if it differs from the next message.
+.
 ## [0.10.0] - 2023-08-28
 
 ### Added